requirements.txt
flask
flask-sqlalchemy
flask-smorest
python-dotenv
marshmallow
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
requests
rq
redis